# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/payments": {
      "post": {
        "operationId": "PaymentsController_create",
        "parameters": [
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent payment creation",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a payment",
        "tags": ["Payments"]
      }
    },
    "/api/v1/payments/{paymentId}/authorize": {
      "post": {
        "operationId": "PaymentsController_authorize",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizePaymentRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Authorize a payment for submission",
        "tags": ["Payments"]
      }
    },
    "/api/v1/payments/{paymentId}/submit": {
      "post": {
        "operationId": "PaymentsController_submit",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitPaymentRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit an authorized payment to the provider",
        "tags": ["Payments"]
      }
    },
    "/api/v1/payments/{paymentId}": {
      "get": {
        "operationId": "PaymentsController_get",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get payment details",
        "tags": ["Payments"]
      }
    },
    "/api/v1/payments/{paymentId}/cancel": {
      "post": {
        "operationId": "PaymentsController_cancel",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelPaymentRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a payment",
        "tags": ["Payments"]
      }
    },
...
    "/api/v1/webhooks/{providerCode}": {
      "post": {
        "operationId": "WebhooksController_handleProviderCallback",
        "parameters": [
          {
            "name": "providerCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Receive provider payment status callbacks",
        "tags": ["Webhooks"]
      }
    }
  },
  "info": {
    "title": "Financial API Platform",
    "description": "Secure financial APIs with OAuth 2.1-aligned authorization patterns, OpenID Connect, and consent management",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "OAuth 2.1 access token"
      },
      "idempotency": {
        "type": "apiKey",
        "in": "header",
        "name": "x-idempotency-key",
        "description": "Idempotency key for safe payment retries"
      },
      "correlation": {
        "type": "apiKey",
        "in": "header",
        "name": "x-correlation-id",
        "description": "End-to-end request correlation identifier"
      }
    },
    "schemas": {
      "CreatePaymentRequestDto": {
        "type": "object",
        "properties": {
          "sourceAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "amountMinor": {
            "type": "number",
            "minimum": 1,
            "example": 1250,
            "description": "Amount in minor units (e.g. cents)"
          },
          "currency": {
            "type": "string",
            "example": "USD",
            "enum": ["USD", "EUR", "GBP", "CAD", "AUD", "JPY"]
          },
          "creditorName": {
            "type": "string",
            "minLength": 1,
            "example": "Acme Utilities"
          },
          "creditorAccountRef": {
            "type": "string",
            "minLength": 1,
            "example": "GB29NWBK60161331926819"
          },
          "reference": {
            "type": "string",
            "minLength": 1,
            "example": "Invoice 1042"
          }
        },
        "required": [
          "sourceAccountId",
          "amountMinor",
          "currency",
          "creditorName",
          "creditorAccountRef",
          "reference"
        ]
      },
      "AuthorizePaymentRequestDto": {
        "type": "object",
        "properties": {
          "actorUserId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["actorUserId"]
      },
      "SubmitPaymentRequestDto": {
        "type": "object",
        "properties": {
          "actorClientId": {
            "type": "string",
            "description": "Defaults to authenticated client"
          }
        }
      },
      "CancelPaymentRequestDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
...
    }
  }
}